Event for opening a file setup
