+++ classes/command_menus.py
@@ -1,0 +1,29 @@
+from typing import List
+from PyQt5.QtWidgets import QMenu, QWidget
+from PyQt5.QtCore import QCoreApplication
+from classes.commands import Command, file_commands, edit_commands
+
+class CommandMenu(QMenu):
+    def __init__(self, parent: QWidget):
+        super().__init__(parent)
+
+    def setMenuCommands(self, cmds: List[Command]):
+        for cmd in cmds:
+            self.addAction(cmd(self))
+        
+    def customEvent(self, event):
+        QCoreApplication.sendEvent(self.parent(), event)
+
+class FileMenu(CommandMenu):
+    def __init__(self, parent: QWidget):
+        super().__init__(parent)
+        self.setTitle("File")
+        self.setMenuCommands(file_commands)
+      
+class EditMenu(CommandMenu):
+    def __init__(self, parent: QWidget):
+        super().__init__(parent)
+        self.setTitle("Edit")
+        self.setMenuCommands(edit_commands)
+  
+menus = [FileMenu, EditMenu]
