+++ main.py
@@ -1,5 +1,7 @@
-from PyQt5.QtWidgets import QApplication, QMainWindow
+from PyQt5.QtWidgets import QApplication, QMainWindow, QLabel, QFileDialog, QDockWidget
+from PyQt5.QtGui import QPixmap
 from classes.controls_bar import ControlsBar
+from PyQt5.QtCore import QEvent
 from sys import argv
 
 class MainWindow(QMainWindow):
@@ -7,6 +9,9 @@ class MainWindow(QMainWindow):
         super().__init__()
         self.setWindowTitle("Pycture")
         self.setMenuBar(ControlsBar(self))
+    
+    def customEvent(self, e):
+        print("Event reached MainWindow")
 
 app = QApplication(argv)
 window = MainWindow()
